Report changes and the last inserted row ID separately from a write
`execute(_:)` returned the changed row count, so a single-expression write
closure inferred its result as `Int` and every call site had to discard it.
It now returns nothing, and a write transaction or connection reports
`changesCount` and `lastInsertedRowID` as properties of the connection it
borrows, which is also where the count and rowid really live.

The `changes` entry point moves from `sqlite3_changes` to
`sqlite3_changes64`, so the count is not truncated by a statement that
changes more rows than an `Int32` holds.

+++ Sources/SQLiteOrbit/SQLite/SQLiteLibrary.swift
@@ -256,8 +256,8 @@ extension SQLiteLibrary {
     public var setBusyTimeout: @Sendable (OpaquePointer?, Int32) -> Int32
     /// Interrupts the query running on a connection: `sqlite3_interrupt`.
     public var interrupt: @Sendable (OpaquePointer?) -> Void
-    /// Rows changed by the most recent statement: `sqlite3_changes`.
-    public var changes: @Sendable (OpaquePointer?) -> Int32
+    /// Rows changed by the most recent statement: `sqlite3_changes64`.
+    public var changes: @Sendable (OpaquePointer?) -> Int64
     /// The rowid of the most recent successful insert: `sqlite3_last_insert_rowid`.
     public var lastInsertedRowID: @Sendable (OpaquePointer?) -> Int64
     /// Whether the connection currently has no transaction open: `sqlite3_get_autocommit`.
@@ -275,7 +275,7 @@ extension SQLiteLibrary {
       setExtendedResultCodes: @escaping @Sendable (OpaquePointer?, Int32) -> Int32,
       setBusyTimeout: @escaping @Sendable (OpaquePointer?, Int32) -> Int32,
       interrupt: @escaping @Sendable (OpaquePointer?) -> Void,
-      changes: @escaping @Sendable (OpaquePointer?) -> Int32,
+      changes: @escaping @Sendable (OpaquePointer?) -> Int64,
       lastInsertedRowID: @escaping @Sendable (OpaquePointer?) -> Int64,
       isAutocommit: @escaping @Sendable (OpaquePointer?) -> Int32
     ) {
